requirements.txt
absl-py==1.3.0
addict==2.4.0
black==22.10.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.12.07
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.5
cox==0.1.post3
cryptography==38.0.3
cycler==0.11.0
Cython==0.29.32
dill==0.3.5.1
dm-control==1.0.3.post1
dm-env==1.5
dm-tree==0.1.7
fancy-gym @ git+https://github.com/dhimiter49/fancy_gym.git@master
fasteners==0.18
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
glfw==2.5.5
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.49.1
gym==0.24.1
gym-notices==0.0.8
h5py==3.7.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.0.0
intel-openmp==2022.2.0
kiwisolver==1.4.4
labmaze==1.0.5
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
MP-PyTorch @ git+https://github.com/ALRhub/MP_PyTorch.git@main
mujoco==2.2.0
mujoco-py==2.1.2.14
mypy-extensions==0.4.3
numexpr==2.8.3
numpy==1.23.5
oauthlib==3.2.2
packaging==21.3
pandas==1.5.0
pathspec==0.10.1
Pillow==9.3.0
pip==22.2.2
platformdirs==2.5.2
protobuf==3.20.1
psutil==5.9.2
py3nvml==0.2.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.0
PyOpenGL==3.1.6
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-git==2018.2.1
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
Send2Trash==1.8.0
setuptools==63.4.1
six==1.16.0
smmap==5.0.0
tables==3.7.0
tbb==2021.7.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wheel==0.37.1
xmltodict==0.13.0
zipp==3.9.0
